pyproject.toml
[project]
name = "flare"
version = "0.0.1"
description = "FLARE Index research prototype — Forward-Lit Authorized Retrieval over Encrypted Indexes"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.8.0",
    "cryptography>=42.0",
    "numpy>=1.26",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "pydantic>=2.6",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
